Fix cocotb tests and eliminate NumPy DeprecationWarnings
Cocotb tests were failing with 'No module named tests' because:
1. tests/__init__.py was missing — needed for cocotb subprocess to import
   tests.cocotb_tests.* modules
2. tests/ref_models/__init__.py was missing — needed for testbench imports
3. Regular test files used 'from ref_models.X' while cocotb testbenches used
   'from tests.ref_models.X' — inconsistent and incompatible

Fixes:
- Create tests/__init__.py and tests/ref_models/__init__.py
- Change all 16 regular test files to use 'from tests.ref_models.X' for
  consistency with cocotb testbench imports
- Fix np.int8() DeprecationWarnings by converting via np.uint8() first:
  test_transformer_block.py:76, test_softmax.py:19,
  test_stateful_softmax.py:19

Results:
- pytest suite: 1234 passed, 0 DeprecationWarnings (45s with -n auto)
- cocotb tests: softmax, alu, gemm all pass (import issue resolved)

+++ tests/test_stateful_softmax.py
@@ -4,7 +4,7 @@ import pytest
 
 from ip_cores.axi_stream_base import AXI4StreamLiteBase
 from ip_cores.stateful_softmax import StatefulSoftmaxCore
-from ref_models.softmax_ref import softmax_ref
+from tests.ref_models.softmax_ref import softmax_ref
 
 
 def _pack_bytes(values: np.ndarray) -> int:
@@ -16,7 +16,7 @@ def _pack_bytes(values: np.ndarray) -> int:
 
 def _unpack_bytes(value: int, T_seq: int) -> np.ndarray:
     return np.array(
-        [np.int8((value >> (i * 8)) & 0xFF) for i in range(T_seq)],
+        [np.int8(np.uint8((value >> (i * 8)) & 0xFF)) for i in range(T_seq)],
         dtype=np.int8,
     )
